multi_agent_pm/requirements.txt
google-adk>=0.5.0
pydantic>=2.0.0
python-dateutil>=2.8.2
fastapi>=0.103.1
uvicorn>=0.23.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
